Added cookie-support to the controller
