+++ server/src/global/auth/authRoutes.ts
@@ -1,0 +1,8 @@
+import express = require("express");
+const router = express.Router();
+
+import * as authController from "./authController";
+
+router.post("/login", authController.login);
+
+module.exports = router;
