+++ server/src/members/membersController.ts
@@ -2,55 +2,10 @@
  * Definition of the handler functions for the members module
  */
 import database = require("../database");
-import auth = require("../global/auth/authController");
 import bcrypt = require("bcryptjs");
 
 import { Request, Response } from "express";
 import * as membersTypes from "./membersTypes";
-import { JWTPayload } from "../global/globalTypes";
-
-/**
- * Obtains username and corresponding permissions
- */
-export const login = (req: Request, res: Response): void => {
-  if (req.body.username === "" || req.body.password === "") {
-    res.status(401).send("Credentials incomplete");
-  } else {
-    database.query(
-      `SELECT mitgliedID, name, passwordHash, GROUP_CONCAT(mitglied_has_berechtigung.berechtigung_berechtigungID) AS permissions
-    FROM mitglied
-    LEFT JOIN mitglied_has_berechtigung ON mitglied.mitgliedID = mitglied_has_berechtigung.mitglied_mitgliedID
-    WHERE mitglied.name = ?
-    GROUP BY mitgliedID, name`,
-      [req.body.username])
-      .then((result: membersTypes.LoginQueryResult[]) => {
-        if (result.length === 0) {
-          res.status(401).send("Username or password wrong");
-        }
-        bcrypt.compare(req.body.password, result[0].passwordHash)
-          .then((match) => {
-            if (match) {
-              const payload: JWTPayload = {
-                mitgliedID: result[0].mitgliedID,
-                name: result[0].name,
-                permissions: result[0].permissions ? result[0].permissions.split(",").map(Number) : []
-              };
-              res.status(200).json({
-                token: auth.generateJWT(payload),
-              });
-            } else {
-              res.status(401).send("Username or password wrong");
-            }
-          })
-          .catch((err) => {
-            res.status(401).send("Username or password wrong");
-          });
-      })
-      .catch((err) => {
-        res.status(500).send("Query Error");
-      });
-  }
-};
 
 /**
  * User can change his password knowing his old one
