+++ server/src/members/membersRoutes.ts
@@ -7,8 +7,6 @@ const router = express.Router();
 import * as authController from "../global/auth/authController";
 import * as membersController from "./membersController";
 
-router.post("/login", membersController.login);
-
 /**
  * =======>>> ALL routes after this point are accessible for loged in users only <<<=======
  */
